IBM-Swift/Kitura#435 Refactored common code into a class and platform specific code into extensions
