+++ Sources/KituraNet/LinuxIncomingHTTPSocketHandler.swift
@@ -23,65 +23,22 @@ import Glibc
 import LoggerAPI
 import Socket
     
-class IncomingHTTPSocketHandler: IncomingSocketHandler {
     
-    private let socket: Socket
-    
-    private weak var delegate: ServerDelegate?
-    
-    ///
-    /// The file descriptor of the incoming socket
-    ///
-    var fileDescriptor: Int32 { return socket.socketfd }
-    
-    private let reader: PseudoAsynchronousReader
-    
-    private let request: HTTPServerRequest
-    
-    // Note: This var is optional to enable it to be constructed in the init function
-    private var response: HTTPServerResponse?
+///
+/// Add Linux specific functionality to the IncomingHTTPSocketHandler. In particular
+/// add the code that handles reading from the socket, writing to the socket, and
+/// the closing of the socket
+extension IncomingHTTPSocketHandler {
     
     ///
-    /// Keep alive timeout in seconds
+    /// Perform platform specfic setup, invoked by the init function
     ///
-    static let keepAliveTimeout: NSTimeInterval = 60
+    func setup() { }
     
-    private(set) var clientRequestedKeepAlive = false
-    
-    private(set) var numberOfRequests = 20
-    
-    var isKeepAlive: Bool { return clientRequestedKeepAlive && numberOfRequests > 0 }
-    
-    enum State {
-        case reset, initial, parsed
-    }
-    
-    private(set) var state = State.initial
-    
-    init(socket: Socket, using: ServerDelegate) {
-        self.socket = socket
-        delegate = using
-        reader = PseudoAsynchronousReader(clientSocket: socket)
-        request = HTTPServerRequest(reader: reader)
-        response = HTTPServerResponse(handler: self)
-    }
-    
-    func process() {
-        switch(state) {
-            case .reset:
-                request.reset()
-                response!.reset()
-                fallthrough
-            case .initial:
-                processInitialState()
-            
-            case .parsed:
-                readAvailableData()
-                break
-        }
-    }
-    
-    private func processInitialState() {
+    ///
+    /// Read in the available data and hand off to ommon processing code
+    ///
+    func handleRead() {
         let buffer = NSMutableData()
         
         do {
@@ -90,9 +47,7 @@ class IncomingHTTPSocketHandler: IncomingSocketHandler {
                 length = try socket.read(into: buffer)
             }
             if  buffer.length > 0  {
-                HTTPServer.clientHandlerQueue.enqueueAsynchronously() { [unowned self] in
-                    self.parse(buffer)
-                }
+                process(buffer, parsingAsynchronously: true)
             }
             else {
                 if  errno != EAGAIN  &&  errno != EWOULDBLOCK  {
@@ -107,63 +62,6 @@ class IncomingHTTPSocketHandler: IncomingSocketHandler {
         }
     }
     
-    private func parse(_ buffer: NSData) {
-        let (parsingState, _) = request.parse(buffer)
-        switch(parsingState) {
-            case .error:
-                break
-            case .initial:
-                break
-            case .headersComplete, .messageComplete:
-                clientRequestedKeepAlive = false
-                parsingComplete()
-            case .headersCompleteKeepAlive, .messageCompleteKeepAlive:
-                clientRequestedKeepAlive = true
-                parsingComplete()
-            case .reset:
-                break
-        }
-    }
-    
-    private func parsingComplete() {
-        state = .parsed
-        delegate?.handle(request: request, response: response!)
-    }
-    
-    func readAvailableData() {
-        let buffer = NSMutableData()
-        
-        do {
-            var length = 1
-            while  length > 0  {
-                length = try socket.read(into: buffer)
-            }
-            if  buffer.length != 0  {
-                reader.addToAvailableData(from: buffer)
-            }
-            else {
-                close()
-            }
-        }
-        catch let error as Socket.Error {
-            Log.error(error.description)
-        } catch {
-            Log.error("Unexpected error...")
-        }
-        
-        return
-    }
-
-    
-    func keepAlive() {
-        state = .reset
-        numberOfRequests -= 1
-    }
-    
-    func drain() {
-        request.drain()
-    }
-    
     ///
     /// Write data to the socket
     ///
@@ -174,25 +72,20 @@ class IncomingHTTPSocketHandler: IncomingSocketHandler {
             try socket.write(from: data)
         }
         catch {
-            print("Write to socket (file descriptor \(socket.socketfd) failed. Error number=\(errno). Message=\(lastError()).")
-            Log.error("Write to socket (file descriptor \(socket.socketfd) failed. Error number=\(errno). Message=\(lastError()).")
+            print("Write to socket (file descriptor \(socket.socketfd) failed. Error number=\(errno). Message=\(errorString(error: errno)).")
+            Log.error("Write to socket (file descriptor \(socket.socketfd) failed. Error number=\(errno). Message=\(errorString(error: errno)).")
         }
     }
     
+    ///
+    /// Close the socket and mark this handler as no longer in progress.
+    ///
     func close() {
         if  socket.socketfd > -1 {
             socket.close()
         }
-    }
-    
-    ///
-    /// Private method to return the last error based on the value of errno.
-    ///
-    /// - Returns: String containing relevant text about the error.
-    ///
-    private func lastError() -> String {
-        
-        return String(validatingUTF8: strerror(errno)) ?? "Error: \(errno)"
+        inProgress = false
+        keepAliveUntil = 0.0
     }
 }
